Fixed that the plugin stops working when adding another classes
It seems that it is unclear on which class Xcode calls pluginDidLoad.
Because of that the plugin stopped working unexpectedly.
To fix that I extended NSObject and moved pluginDidLoad there. I did
the fix for both objective c and swift. I described my solution also
here http://stackoverflow.com/a/30000663/1442208

This closes #31, and #32.